Add support for external references
Enable external refs with -ri

External references will be encapsulated in a package as described by the
--extrefs argument.

Get can be insecure (eg ignore Cert issues) using the -insecure option

Does not generate the external code (as it would be in a different file anyway)
but allows the code to be referenced.  Automatically adds import paths/package
name to appropriate generated files.