Attempting to fix date query param parsing
Because new types are created for all fields of top-level param structs,
it is difficult to reason about how to handle our custom types.

Here I am adding an additional field to the Date type that will
(hopefully) allow us to identify all types that might be wrapping it
via reflect.ConvertibleTo.