fix: fix panic when printing the nil Date
Not sure why, but this is only happening when running with delve.
It seems like printing a nil Stringer causes a panic when debugging
with delve but is handled gracefully outside of delve.

Steps to reproduce:
1. Start debugger
2. Print the value of a nil Date
```go
var d *Date
fmt.Sprintf("%v", d)
```

Theory:
`time.Time` is embedded within the `Date` type. This means that if
Date is nil and Stringer is used to get the string representation of
it, it'll try to call Date.Time.String() which will result in a nil
dereferencing error.