feat: Add x-oapi-codegen-middlewares for paths and operations
This change further extends the availability of the
x-oapi-codegen-middleware tag to be used for both specific operations,
and also for a group of tags, helping slightly reduce duplication within
schema creation.

The order for middlewares is defined as being global middleware first,
then path middleware, and then particular operation-specific middleware.