Fix customProvider RequestEditorFn example
The ctx and req arguments in the example appear to be ordered incorrectly.
This commit modifies the example to list ctx first, which appears to be the
correct ordering.

I'm just learning golang so I haven't dug deeply to confirm this. Just noted
that the example wasn't working as-is in my test project, but compiles
when I invert them, and ctx documentation says it should usually come first.