Allow custom JSON media types to be marked as JSON
As noted in #581, we won't currently support custom media types using
the suffixes, such as custom or vendor media types.

This also adds coverage for the OpenAPI spec's requirement for only a
single MediaType to be defined.

We also add an example to the generated client, to ensure that a
vendored JSON type can be generated.

Closes #581.