Generate type names in GoDoc comments
To make sure that our commented types appear as "official" GoDoc
comments, we need to prefix each

To avoid duplicating too much code, we can create an overloaded
`StringWithTypeNameToGoComment` which delegates to a shared
`stringToGoCommentWithPrefix`.

We then need to move this execution into the template, as we do not have
access to the `TypeName` at the point prior that we're attempting to
generate our comments.

This also includes the fields that are generated inside structs.

This includes any `make generate` changes, too.