Add basic automated builds
To ensure that contributions are well-formed, we can set up basic CI.

We can do this following [0] as inspiration, and using our `make` tasks
to automate the process.

We can also create a separate workflow for validating that the `make
generate` has been run, and no untracked changes are present, as well as
that `go mod tidy` has been run.

[0]: https://blog.kowalczyk.info/article/8dd9c2c0413047c589a321b1ccba7129/using-github-actions-with-go.html
[1]: https://www.jvt.me/posts/2022/04/29/git-uncommitted-changes/