Allow binding types as parameters through `String()` methods
If trying to use a type, such as a UUID in a path parameter, we'd
previously receive an error at runtime to say that `types.UUID` isn't
supported.

As this is a common use case, we should support it. Instead of adding
support for it manually, we can instead look at whether the there's a
`String` method, and if so use that to serialise the field.