Add functionality to merge union types
As noted in #651, it's currently not possible to merge in a data type,
which means when we want to use an `allOf`, we end up only being able to
add a single type.

This introduces a new `Merge` method on generated union types, which
delegates to our `runtime.JsonMerge` method, which uses the `jsonmerge`
library.

We need to handle the case that a slice may be `nil`, and that the
upstream library does not have support for JSON arrays.

Closes #651.