Remove generation of code to handle optional query parameters
As highlighted in #658, in the case that a query parameter is optional,
and is not JSON, a passthrough type or a styled parameter, we are
rendering an empty method.

To simplify our generated code, we can only generate it as and when it's
required.

To make this more readable, we can indent this new optional block in the
template.

Closes #658.