Add fallback for old configuration style
As a means to avoid the breaking changes we introduced in v1.11.0, where
the new-style configuration was made the default, we can tweak our
heuristic for how we work out what configuration is in use by only using
the new-style config if the new `--config` flag is used.

This change in itself is non-breaking, and makes it possible to continue
using the old configuration arguments, as long as you don't provide the
new `--config` flag.

Similar to thoughts in #629, but does not require a revert.

Closes #629.