Remove io/ioutil
This Commit edits the go templates so that io/ioutil is replaced by io.

This will make oapi-codegen generate code that is compliant with [staticcheck](https://staticcheck.io/).

Since go 1.16 io/ioutil is deprecated. The current version of oapi-codegen
requires go 1.16 to compile. Most users will also use at least go 1.16
to run the generated code because no binaries are distributed, but users
need to compile oapi-codegen themselves.
If not, this is would be a breaking change.

Signed-off-by: leonnicolas <leonloechner@gmx.de>