REW!!!! Generate anonymous objects referenced in `paths`
Previously, there was no way to render any anonymous types that are
referenced in the `paths` of an OpenAPI specification.

Although there is a strong preference in the community to use the
`/components/schemas/`, there are times that it's out of our control to
move objects around.

This generates both the `requestBody` and `responses`.

TODO Closes 626 TODO