Generate raw types for non-JSON `responses`
As part of future changes we need to be able to reference all the types
in the defined `responses` of a specification.

However, as we're currently only rendering `application/json` media
types, this fails.

Instead, we can generate `json.RawMessage` (which is an alias of
`[]byte`) for these types, which allow referencing them, and
(un)marshaling data to the types.