Improve logic for x-omitempty
Currently x-omitempty only enables you to force the _absences_ of `omitempty`.  This PR changes the logic so that the existance of `x-omitempty` will force omit empty to be whatever the user specifies.

The logic around omit empty is pretty confusing here though.  For example, why is omitEmpty excluded by default if it's nullable (e.g. if `p.Nullable` then `omitEmpty = false`).