Fix: Allow multiple JSON types to be generated
As flagged in #1127, the use of `IsMediaTypeJson` leads to potentially
having duplicate types generated, leading to compilation errors.

We can instead only generate i.e. `CreateWhateverJSONRequestBody` for
the `application/json` type, and then construct a specific type for each
media type with a matching name.

This requires we add a new utility method `IsJSON` to simplify our
checks, and wire it in in each place that requires we render
JSON-specific code.

Closes #1127.