Adds SAML and SCIM to All Caps List
Adds SAML and SCIM to All Caps List. SAML is an XML standard that allows secure web domains to exchange user authentication and authorization data. SCIM is a standard for automating the exchange of user identity information between identity domains, or IT systems. Both SAML and SCIM are fairly common acronyms used in many code-bases today.