Fix: Generate models for all JSON media types
As highlighted in #1168, we have some cases where JSON-compatible media
types aren't having their respective models generated, as we previously
only looked for `application/json`.

To resolve this, we can make sure that any media types that are
compatible with JSON (via our utility method) can have their types
generated.

This is a fix for #1168, which could also be deemed a feature. However,
as this has broken in the upgrade from v1.12.x to v1.13.x, this is being
treated as a feature.

Closes #1168.