Fix: Allow consumers to specify how many circular references are allowed
As noted in #936, we have an issue where a number of circular references
may be expected and allowed, so we should provide the ability to
configure this, allowing users to decide what makes sense for them.

This requires we add an additional method,
`LoadSwaggerWithCircularReferenceCount`, as the existing method is part
of our public API and we cannot modify it without breaking consumers.

This is already available in the version of kin-openapi we're using,
we're just configuring it.

Closes #936, #957.