Migrate to oapi-codegen/runtime package
As part of #1142 we can start to reduce our transitive dependencies
by refactoring the `pkg/runtime` as it exists in
deepmap/oapi-codegen.

This allows us to introduce a breaking change (tweaking the structure of
this package, including moving the `strictmiddleware` concepts) by
creating a fresh package. This does require we tweak the way that our
imports and generated code works.

By moving `runtime` to a separate package, we can allow more frequent
updates to the runtime-specific code, without tying it directly to the
version of the code generator in use.