Bump fiber to v2.49.1
As part of the upgrade to Go 1.21, the Go toolchain now requires the
`go` directive to match the maximum Go version in use in dependencies.

This leads to any transitive dependency on fiber to result in a
requirement of the consuming project moving to Go 1.21.

We can upgrade fiber to a version that has downgraded the Go 1.21
requirement.

As part of #1221.