Bump Iris to remove Go 1.21 dependency
As part of the upgrade to Go 1.21, the Go toolchain now requires the
`go` directive to match the maximum Go version in use in dependencies.

This leads to any transitive dependency on Iris to result in a
requirement of the consuming project moving to Go 1.21.

We can upgrade Iris to a version that has downgraded the Go 1.21
requirement.

We will pin to the latest commit on HEAD until v12.2.6 is released.

Closes #1221.