Allow disabling type aliases for arrays
As noted in #1247, in some cases it can be useful to have arrays defined
as a concrete type, not a type alias, so you are able to extend them
with methods.

This produces a new opt-in output-option,
`disable-type-aliases-for-type`, which can be used to control this
behaviour

Right now, we only want to add support for `array`s, but the
configuration is extensible in the future.

This also introduces a test case to demonstrate the capabilities.

We need to introduce a `sliceContains` method to simplify this lookup,
as we don't yet have `slices.Contains` due to the Go versions we're
targeting.