Migrate middleware to their own Go modules
Similar to our work with the migration to separate modules for the
`pkg/runtime` and `pkg/types` packages, we can do the same for the
middleware.

This allows us to iterate over these quicker than the underlying code
generator and gives us the opportunity to introduce breaking changes
more easily if appropriate.

These won't be removed until the next major version, but we can mark
them as deprecated to nudge folks to move away from them.