Update echo params to be idiomatic
Echo uses `c` as the `echo.Context` parameter name, differentiating it from the widely used `ctx` for a `context.Context`.

This commit updates the generated code to use the `c` convention to bring it in line with Echo's common practice and make generated code more readable.

This change will not break existing code.  Existing `ServerInterface` do not need to have the same parameter names, so all existing code that uses `ctx echo.Context` will be able to ignore the change (or update when it's convenient).