Support media type parameters for `IsMediaTypeJson`
Although not _technically_ a valid extension to JSON according to [the
media type definition], it's got a number of services using it for a `v`
or `version` parameter, so we should support this as a valid option.

This makes sure that we parse it as a valid media type, performing the
same checks as we did before by ignoring the parameters.

[the media type definition]: https://www.iana.org/assignments/media-types/application/json