build: create a `Makefile` per module
As part of future changes to introduce the `net/http`-only server, we'll
introduce the reliance on Go 1.22-only code.

The way that we currently lint/build/etc the project doesn't make this
easy to do so, so we should instead refactor our build steps to make
this easier.

This is the best option I can see right now, but will be happy to tweak
it if we can find an alternative.