docs: completely rewrite the README
I've been thinking for some time we should rethink the documentation
we've got, as it's been feeling a little bit underdocumented, a bit of a
mixed voice, and we've also had not-that-actionable examples for all the
functionality.

This is the culmination of ~7 weeks of rewriting, alongside the addition
of new examples.

Closes #1254.

Co-authored-by: Jamie Tanna <jamie.tanna@elastic.co>