feat: add JSON schema for configuration files
To make it easier for our users to understand the configuration options
that they're able to use, we can introduce a JSON schema that can be
used by an IDE or Language Server Protocol (LSP) implementation to
validate + provide autocomplete.

We make sure to use `additionalProperties: false` to make it super clear
when there's no additional properties possible, which will also make it
clearer when we have items in our configuration which may not be valid,
or from a newer version of `oapi-codegen` that's in use.

Closes #1546.