fix: ensure we pre-import `nullable`
There was a report (via Twitter) that `nullable` wasn't getting
generated for them.

It appears we may not be seeing the issue due to us already having an
import of `nullable`, and therefore when `goimports` runs, it knows the
library is here.

Instead, we can make sure that we pre-import it for folks who don't
already have it in their `go.mod`s.