feat(extension): add `x-oapi-codegen-only-honour-go-name`
Although originally to fix #1638, which isn't a very recommended
approach as it may lead to a confusing behaviour for users, we can still
add support.

However, we should hide this behind both a global Compatibility Option,
to ensure that it's clearly opted-in, as well as adding a per-field
level to ensure that only the right fields are overridden.

Closes #1638.