Reject parsing non-compliant OpenAPI specs
This kinda fixes #867 by rejecting specs that we know
we can't parse.

We only support 3.0.x openAPI specs, which must have the
`openapi` property as semver at the top level.

Exit with an error if `openapi` is not present. This will be
the case for Swagger specs.

Exit with an error if major version != 3

Exit with an error if minor version != 0

Stop using the word Swagger in function names where we can help it.