Fix #867 by intercepting spec loading
This change adds a shim into the Kin loader which attempts to find
an `openapi` or `swagger` version key in the spec, and returns an
error if it's an unsupported version. in the case where the version
can't be detected, we do nothing and behave as previously, failing
in the parsing step.