build: require Go 1.21 to build
As part of future changes, we want to introduce a dependency on a
library that requires Go 1.21, therefore we can preemptively bump the
requirement.

This also is not too controversial a requirement, as Go 1.21 has just
gone end-of-life, so consumers are ideally on Go 1.22, the minimum
supported Go version.

We make sure to pin to the latest patch release of Go 1.21.