feat: add support for OpenAPI Overlay
In the past, users wishing to override specific configuration, for
instance taking advantage of extensions such as `x-go-type` would need
to modify the OpenAPI specification they are using.

In a lot of cases, this OpenAPI specification would be produced by a
different team to the consumers (or even a different company) and so
asking them to make changes like this were unreasonable.

This would lead to the API consumers needing to vendor the specification
from the producer (which is our recommendation anyway) and then make any
number of local changes to the specification to make it generate code
that looks reasonable.

However, in the case that a consumer would update their specification,
they would likely end up with a number of merge conflicts.

With these changes, it is now possible to make changes to the input
OpenAPI specification _without needing to modify it directly_.

This takes advantage of the OpenAPI Overlay specification[0], which is a
stable specification, and is ready for wider implementation.

This uses Speakeasy's implementation, which is noted as an "alpha"
implementation.

We introduce a new function, `LoadSwaggerWithOverlay`, which performs
the wrangling of the Overlay on top of the specification we're pointing
to, and make sure we wire in the ability to configure the path, as well
as whether we want to relax the "strict" validation the Overlay library
provides.

Closes #1553.

[0]: https://github.com/OAI/Overlay-Specification