docs(sponsors): clarify Elastic's 4h/mo
This makes it a little bit clearer that Elastic's 4h/mo of in-work
support that I'm able to do is different to the GitHub Sponsor'd support
we receive from others.

This isn't something that other sponsors can do right now, so we should
make it clearer that Elastic's at a higher level.

But we are _super appreciative_ of all of our sponsors!