fix(codegen): allow exposing the input spec's `operationId`
As noted in #1274, there appears to be a modification of the underlying
`operationId` in a given Operation, which is leading to the Embedded
Spec getting a different `operationId` than our input specification.

This appears to be due to the reference to the `Operation` being
modified (unintentionally, it would seem) in `OperationDefinitions`,
resulting in a `nameNormalizer`'d `operationId` being emitted, instead
of the input specification's `operationId`.

We can make sure that this is documented behaviour for the future, for
anyone else modifying this file.

To allow users to modify this, we can introduce an opt-in Compatibility
flag, `preserve-original-operation-id-casing-in-embedded-spec`.

Although this would be ideally an opt-out, this has the risk of breaking
existing code, so we can't push it to users.

To make sure that the underlying `operationId` isn't (always) modified,
we can take a copy of the value and then only update it if it's not
opted-out.

This also wires in a new test-case to document behaviour and to catch
issues in the future.

Closes #1274.