fix(codegen): allow using x-go-type and x-go-type-skip-optional-pointer together
When using `x-go-type`, `GenerateGoSchema` returns early, meaning that
any extensions that are checked after this one are no-op.

This PR moves the code for `x-go-type-skip-optional-pointer` higher up,
so it can be used alongside `x-go-type`.

This corrects this piece of behaviour, and amends our generated code
examples to correctly apply `x-go-type-skip-optional-pointer` when using
an optional field, or inside an `AllOf`.