docs(std-http-server): warn when not using Go 1.22
Although this is less necessary after bumping our minimum Go version to
1.22 (in 0e8156a14e1a92e59194230873b2053d5b29e38b), it's still
worthwhile ensuring that the module we're generating the code for is
correctly set to Go 1.22, otherwise users will receive:

    HTTP/1.1 404 Not Found
    ...
    404 page not found

To do this, we can introduce a new means to flag "warnings" that need to
be output.

We need to make sure we're using `ParseLax` to reduce the risk of new
directives or content being introduced into a `go.mod` breaking
`oapi-codegen`.

This is a step towards being able to implement future warnings to ensure
that consumers are using correct Go versions.

Closes #1628.

[0]: https://www.jvt.me/posts/2024/03/04/go-net-http-why-404/