feat(output-options): add `prefer-skip-optional-pointer-on-container-types`
One of the key things `oapi-codegen` does is to use an "optional
pointer", following idiomatic Go practices, to indicate that a
field/type is optional.

It's possible for users to override this using behaviour on a per-type
or per-field basis using the `x-go-type-skip-optional-pointer`
extension, but there are cases, as noted in #1561, where we're
generating the optional pointer for container types (such as slices or
maps) which doesn't make sense, as it requires an extra level of
`... != nil` checks.

Therefore we should introduce a new opt-in Output Option to allow
skipping the optional pointer on a container type.

Closes #1561.