feat(output-options): add `prefer-skip-optional-pointer` to default to skipping optional pointers
One of the key things `oapi-codegen` does is to use an "optional
pointer", following idiomatic Go practices, to indicate that a
field/type is optional.

It's possible for users to override this using behaviour on a per-type
or per-field basis using the `x-go-type-skip-optional-pointer`
extension, but as has been flagged before (#1454, #1860), this can be
very cumbersome.

It's possible to use an OpenAPI Overlay to simplify this - but comes at
the cost of likely using a complex JSON Path - so we should provide a
more reasonable option to default this behaviour, with an Output Option.

This will be an opt-in default, but with the ability to override the
behaviour on a per-type/per-field basis, allowing control of when we
should be able to use them.

This also introduces a test to validate this behaviour.

Co-authored-by: Jamie Tanna <jamie@jamietanna.co.uk>