feat(templates): add `HasOptionalPointer` helper
As a means to consolidate our logic for whether a type has an optional
pointer in templates, we can introduce a new helper method,
`HasOptionalPointer`, which will indicate whether a given type - in a
template - should have an "optional pointer" to indicate whether the
field is optional.

This wires it in for both properties (of structs) or parameter
definitions.

To start with, we will derive this based on whether it is required, and
will add support for the `x-go-type-skip-optional-pointer` extension
afterwards.