feat(output-options): allow using `omitzero` with `prefer-skip-optional-pointer`
One of the key things `oapi-codegen` does is to use an "optional
pointer", following idiomatic Go practices, to indicate that a
field/type is optional.

As noted in #1899, now Go 1.24+ has the `omitzero` JSON tag for
marshaling only when a type is not its zero value.

As we support generating `omitzero` JSON tags via the `x-omitzero`
extension, we can build on top of this to consider allowing the use of
`omitzero` when using the newly added global Output Option,
`prefer-skip-optional-pointer` to tune this behaviour.

This introduces a new Output Option,
`prefer-skip-optional-pointer-with-omitzero`, which mirrors behaviour
from `prefer-skip-optional-pointer`, but produces an `omitzero` flag for
any skipped pointer types, to not marshal the type if the zero value.

This allows folks using Go 1.24+ to much more ergonomically work with
different (optional) types, while allowing `omitzero` to simplify
(un)marshalling.

Right now, we don't have a straightforward means of enforcing/warning
that `oapi-codegen` is being generated into a non-Go-1.24+ project, so
we'll only document the behaviour.

Closes #1899.