Fix server variables conflict
Adds a property to opt-in avoiding the default value colliding with the
value `default` for enum server variables. When set to true `Enum` is
added to the variable name so that it doesn't clash with the default
variable name. This behaviour is opt-in to avoid backward compatibility,
but the default is actually a bug, so it make sense to change that
default in the next major release.

Resolves: https://github.com/oapi-codegen/oapi-codegen/issues/2003