build: add Scorecards workflow
This is another step towards better understanding this project's health,
and making it clear to consumers.

Although we're already onboarded via the upstream project[0] (via the
"old" repo name), it's better to have this a first-class product of our
project, as this also allows validating i.e. branch protection due to
`permissions: read-all`.

Takes configuration via Renovate's usage[1], alongside the suggested
version from the Scorecard project.

This uploads the SARIF results to show that we have issues in the
Security tab (for maintainers).

[0]: https://securityscorecards.dev/viewer/?uri=github.com%2Fdeepmap%2Foapi-codegen
[1]: https://github.com/renovatebot/renovate/blob/8b86b8cdb4a3e36d6211e47a2e6a201f25f674da/.github/workflows/scorecard.yml