fix: report formatting errors with line numbers
When formatting the resulting output, using `goimports`, we may
receive an error due to generating invalid code using our templates.

The output returned is fairly unhelpful, noting a line number, but
requiring the users to take the broken output and then paste it into
something that has a line number.

This improves the output by reporting:

- line numbers for the resulting Go code
- a ❗ where the (first) formatting error is found