chore(templates): add `Property.ZeroValueIsNil` helper method
As part of #2031, we need to correct a subtle breakage in #1981, where
we would not correctly perform the nil check on arrays, breaking
marshalling functionality.

As a first step, we add a new helper method to inform as to whether the
zero value of a Property could be `nil`, and therefore needs a `nil`
check.