chore(templates): add `Property.RequiresPointerCheck` helper method
As part of #2031, we need to correct a subtle breakage in #1981, where
we would not correctly perform the nil check on arrays, breaking
marshalling functionality.

As a first step, we add a new helper method to be used to validate
whether a `nil` check is required on a given type, which is used for
either types with a zero value of `nil`, or where we have an optional
pointer set.